api: Update triggers API
This has never really been used, so I'm going to break the functionality
here so that the new approach doesn't return the values of the secrets.
There's really no need to ever let secrets leave this system via an API.

This also potentially loosens up the permissions on this API. Since
nothing secret is being shared, anyone who can "build" the project
should be able to see the trigger information.

Signed-off-by: Andy Doan <andy@foundries.io>